Introduce test container adn update draft accounts integ test to use it
