+++ src/integrationTest/java/uk/gov/hmcts/opal/controllers/DraftAccountControllerErrorsIntegrationTest.java
@@ -1,34 +1,19 @@
 package uk.gov.hmcts.opal.controllers;
 
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.persistence.QueryTimeoutException;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.Arguments;
-import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.Mockito;
 import org.mockito.stubbing.OngoingStubbing;
 import org.postgresql.util.PSQLException;
 import org.postgresql.util.PSQLState;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
-import uk.gov.hmcts.opal.authorisation.model.Permissions;
-import uk.gov.hmcts.opal.authorisation.model.UserState;
-import uk.gov.hmcts.opal.config.WebConfig;
-import uk.gov.hmcts.opal.controllers.advice.GlobalExceptionHandler;
-import uk.gov.hmcts.opal.dto.AddDraftAccountRequestDto;
-import uk.gov.hmcts.opal.dto.ReplaceDraftAccountRequestDto;
-import uk.gov.hmcts.opal.dto.ToJsonString;
-import uk.gov.hmcts.opal.dto.UpdateDraftAccountRequestDto;
+import uk.gov.hmcts.opal.AbstractIntegrationTest;
 import uk.gov.hmcts.opal.entity.BusinessUnitEntity;
 import uk.gov.hmcts.opal.entity.DraftAccountEntity;
 import uk.gov.hmcts.opal.service.opal.DraftAccountService;
@@ -37,32 +22,21 @@ import uk.gov.hmcts.opal.service.opal.UserStateService;
 import java.net.ConnectException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Stream;
 
-import static org.hamcrest.Matchers.containsString;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.put;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.hmcts.opal.controllers.util.UserStateUtil.allPermissionsUser;
-import static uk.gov.hmcts.opal.controllers.util.UserStateUtil.noPermissionsUser;
-import static uk.gov.hmcts.opal.controllers.util.UserStateUtil.permissionUser;
 import static uk.gov.hmcts.opal.entity.DraftAccountStatus.SUBMITTED;
 
 
-@WebMvcTest
-@ContextConfiguration(classes = {DraftAccountController.class, GlobalExceptionHandler.class, WebConfig.class})
 @ActiveProfiles({"integration"})
 @Slf4j(topic = "DraftAccountControllerErrorsIntegrationTest")
-class DraftAccountControllerErrorsIntegrationTest {
+class DraftAccountControllerErrorsIntegrationTest extends AbstractIntegrationTest {
     private static final String URL_BASE = "/draft-accounts";
 
     private static final Short BU_ID = (short)007;
@@ -77,207 +51,11 @@ class DraftAccountControllerErrorsIntegrationTest {
     @MockBean
     UserStateService userStateService;
 
-    @Test
-    void testGetDraftAccountById_trap403Response_wrongPermission() throws Exception {
-        DraftAccountEntity entity = createDraftAccountEntity(BU_ID);
-        when(draftAccountService.getDraftAccount(2L)).thenReturn(entity);
-
-        UserState userState = permissionUser(BU_ID, Permissions.COLLECTION_ORDER);
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(userState);
-
-        mockMvc.perform(
-                get(URL_BASE + "/2")
-                    .header("authorization", "Bearer some_value"))
-            .andExpect(status().isForbidden())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.error").value("Forbidden"))
-            .andExpect(jsonPath("$.message").value(
-                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS, CHECK_VALIDATE_DRAFT_ACCOUNTS] "
-                    + "permission(s) are not enabled for the user."));
-    }
-
-    @Test
-    void testGetDraftAccountById_trap403Response_wrongBusinessUnit() throws Exception {
-        DraftAccountEntity entity = createDraftAccountEntity(BU_ID);
-        when(draftAccountService.getDraftAccount(2L)).thenReturn(entity);
-
-        UserState userState = permissionUser((short)005, Permissions.DRAFT_ACCOUNT_PERMISSIONS);
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(userState);
-
-        mockMvc.perform(
-                get(URL_BASE + "/2")
-                    .header("authorization", "Bearer some_value"))
-            .andExpect(status().isForbidden())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.error").value("Forbidden"))
-            .andExpect(jsonPath("$.message").value(
-                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS, CHECK_VALIDATE_DRAFT_ACCOUNTS] "
-                    + "permission(s) are not enabled in business unit: 7"));
-    }
-
-    @Test
-    void testGetDraftAccountById_trap404Response() throws Exception {
-        DraftAccountEntity entity = Mockito.mock(DraftAccountEntity.class);
-        when(entity.getBusinessUnit()).thenThrow(new EntityNotFoundException());
-
-        when(draftAccountService.getDraftAccount(2L)).thenReturn(entity);
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
-
-        mockMvc.perform(
-            get(URL_BASE + "/2")
-                .header("authorization", "Bearer some_value"))
-            .andExpect(status().isNotFound())
-        ;
-    }
-
-    @Test
-    void testGetDraftAccountById_trap406Response() throws Exception {
-        when(draftAccountService.getDraftAccount(1L)).thenReturn(createDraftAccountEntity(BU_ID));
-        shouldReturn406WhenResponseContentTypeNotSupported(get(URL_BASE + "/1"));
-    }
-
-    @Test
-    void testGetDraftAccountById_trap408Response() throws Exception {
-        shouldReturn408WhenTimeout(get(URL_BASE + "/1"),
-                                   when(draftAccountService.getDraftAccount(1L)));
-    }
-
-    @Test
-    void testGetDraftAccountById_trap503Response() throws Exception {
-        shouldReturn503WhenDownstreamServiceIsUnavailable(get(URL_BASE + "/1"),
-                                                          when(draftAccountService.getDraftAccount(1L)));
-    }
-
-    @Test
-    void testGetDraftAccountsSummaries_trap400Response() throws Exception {
-        List<DraftAccountEntity> draftAccountEntities = createDraftAccountEntityList("Dave");
-        final Short businessId = (short)1;
-
-        UserState user = permissionUser(businessId, Permissions.CREATE_MANAGE_DRAFT_ACCOUNTS);
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(user);
-        when(draftAccountService.getDraftAccounts(any(), any(), any(), any()))
-            .thenReturn(draftAccountEntities);
-
-        mockMvc.perform(get(URL_BASE)
-                            .header("authorization", "Bearer some_value")
-                            .param("business_unit", businessId.toString())
-                            .param("submitted_by", "Dave")
-                            .param("not_submitted_by", "Tony")
-                            .contentType(MediaType.APPLICATION_JSON))
-            .andExpect(status().isBadRequest())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.error").value("Bad Request"))
-            .andExpect(jsonPath("$.message")
-                           .value("Cannot include both 'submitted_by' and 'not_submitted_by' parameters."));
-    }
-
-    @Test
-    void testGetDraftAccountsSummaries_trap403Response_noPermission() throws Exception {
-        final Short businessId = (short)1;
-
-        UserState user = permissionUser(businessId, Permissions.COLLECTION_ORDER);
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(user);
-        when(draftAccountService.getDraftAccounts(any(), any(), any(), any()))
-            .thenReturn(createDraftAccountEntityList("Tony"));
-
-        mockMvc.perform(get(URL_BASE)
-                            .header("authorization", "Bearer some_value")
-                            .param("business_unit", businessId.toString())
-                            .contentType(MediaType.APPLICATION_JSON))
-            .andExpect(status().isForbidden())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON));
-    }
-
-    @Test
-    void testGetDraftAccountsSummaries_trap406Response() throws Exception {
-        shouldReturn406WhenResponseContentTypeNotSupported(get(URL_BASE));
-    }
-
-    @Test
-    void testGetDraftAccountsSummaries_trap408Response() throws Exception {
-        shouldReturn408WhenTimeout(get(URL_BASE), when(draftAccountService.getDraftAccounts(any(), any(), any(),
-                                                                                            any())));
-    }
-
-    @Test
-    void testGetDraftAccountsSummaries_trap503Response() throws Exception {
-        shouldReturn503WhenDownstreamServiceIsUnavailable(get(URL_BASE),
-                                             when(draftAccountService.getDraftAccounts(any(), any(), any(),
-                                                                                       any())));
-    }
-
-    private DraftAccountEntity createDraftAccountEntity(short businessUnitId) {
-        return createDraftAccountEntity("Tony", businessUnitId);
-    }
-
-    private DraftAccountEntity createDraftAccountEntity(String submittedBy) {
-        return createDraftAccountEntity(submittedBy, BU_ID);
-    }
-
-    private DraftAccountEntity createDraftAccountEntity(String submittedBy, short businessUnit) {
-        return DraftAccountEntity.builder()
-            .draftAccountId(1L)
-            .businessUnit(BusinessUnitEntity.builder().businessUnitId(businessUnit).build())
-            .createdDate(LocalDate.of(2023, 1, 2).atStartOfDay())
-            .submittedBy(submittedBy)
-            .accountType("DRAFT")
-            .accountStatus(SUBMITTED)
-            .statusMessage("Status is OK")
-            .accountStatusDate(LocalDateTime.of(2024, 11, 11, 11, 11))
-            .account(validAccountJson())
-            .accountSnapshot("{ \"data\": \"something snappy\"}")
-            .timelineData(validTimelineDataJson())
-            .build();
-    }
-
-    private List<DraftAccountEntity> createDraftAccountEntityList(String... submittedBys) {
-        return Arrays.stream(submittedBys).map(this::createDraftAccountEntity).toList();
-    }
-
-    @Test
-    void testReplaceDraftAccount_trap403Response_boPermission() throws Exception {
-        Long draftAccountId = 241L;
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
-
-        mockMvc.perform(put(URL_BASE + "/" + draftAccountId)
-                                               .header("authorization", "Bearer some_value")
-                                               .contentType(MediaType.APPLICATION_JSON)
-                                               .content(validCreateRequestBody()))
-            .andExpect(status().isForbidden())
-            .andReturn();
-
-    }
-
-
-    @Test
-    void testUpdateDraftAccount_trap403Response_noPermission() throws Exception {
-        Long draftAccountId = 241L;
-        String requestBody = "            {\n"
-            + "                \"account_status\": \"PENDING\",\n"
-            + "                \"validated_by\": \"BUUID1\",\n"
-            + "                \"business_unit_id\": 5,\n"
-            + "                \"timeline_data\": "
-            + validTimelineDataJson()
-            + "\n"
-            + "            }";
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
-
-        mockMvc.perform(patch(URL_BASE + "/" + draftAccountId)
-                            .header("authorization", "Bearer some_value")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(requestBody))
-            .andExpect(status().isForbidden())
-            .andReturn();
-
-    }
 
     @Test
     void testUpdateDraftAccount_trap406Response() throws Exception {
-        when(draftAccountService.updateDraftAccount(any(), any())).thenReturn(createDraftAccountEntity(BU_ID));
+        when(draftAccountService.updateDraftAccount(any(), any()))
+            .thenReturn(createDraftAccountEntity("Test",BU_ID));
         shouldReturn406WhenResponseContentTypeNotSupported(
             patch(URL_BASE + "/1").contentType(MediaType.APPLICATION_JSON).content(validUpdateRequestBody())
         );
@@ -300,87 +78,15 @@ class DraftAccountControllerErrorsIntegrationTest {
     }
 
     @Test
-    void testPostDraftAccount_trap400Response() throws Exception {
-
-        String expectedErrorMessageStart =
-            "JSON Schema Validation Error: Validating against JSON schema 'addDraftAccountRequest.json',"
-                + " found 14 validation errors:";
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
-
-        mockMvc.perform(post(URL_BASE)
-                            .header("authorization", "Bearer some_value")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(invalidCreateRequestBody()))
-            .andExpect(status().isBadRequest())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.error").value("Bad Request"))
-            .andExpect(jsonPath("$.message").value(containsString(expectedErrorMessageStart)))
-            .andExpect(jsonPath("$.message").value(containsString("required property 'account_type' not found")))
-            .andExpect(jsonPath("$.message").value(containsString("required property 'submitted_by' not found")))
-            .andExpect(jsonPath("$.message").value(containsString("required property 'submitted_by_name' not found")))
-            .andExpect(jsonPath("$.message").value(containsString("required property 'timeline_data' not found")));
-
-    }
-
-
-
-    @Test
-    void testPostDraftAccount_trap403Response_noPermission() throws Exception {
-
-        String validRequestBody = setupValidPostRequest();
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
-
-        MvcResult result = mockMvc.perform(post(URL_BASE)
-                            .header("authorization", "Bearer some_value")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(validRequestBody))
-            .andExpect(status().isForbidden())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.error").value("Forbidden"))
-            .andExpect(jsonPath("$.message").value(
-                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS] permission(s) are not enabled for the user."))
-            .andReturn();
-
-        String body = result.getResponse().getContentAsString();
-
-        log.info(":testPostDraftAccount_permission: Response body:\n" + ToJsonString.toPrettyJson(body));
-    }
-
-    @Test
-    void testPostDraftAccount_trap403Response_wrongPermission() throws Exception {
-
-        String validRequestBody = setupValidPostRequest();
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(
-            permissionUser((short)5, Permissions.CHECK_VALIDATE_DRAFT_ACCOUNTS, Permissions.ACCOUNT_ENQUIRY));
-
-        MvcResult result = mockMvc.perform(post(URL_BASE)
-                            .header("authorization", "Bearer some_value")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(validRequestBody))
-            .andExpect(status().isForbidden())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.error").value("Forbidden"))
-            .andExpect(jsonPath("$.message").value(
-                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS] permission(s) are not enabled for the user."))
-            .andReturn();
-
-        String body = result.getResponse().getContentAsString();
-
-        log.info(":testPostDraftAccount_permission: Response body:\n" + ToJsonString.toPrettyJson(body));
-    }
-
-    @Test
     void testPostDraftAccount_trap406Response() throws Exception {
-        String validRequestBody = setupValidPostRequest();
+        String validRequestBody = validCreateRequestBody();
         shouldReturn406WhenResponseContentTypeNotSupported(
             post(URL_BASE).contentType(MediaType.APPLICATION_JSON).content(validRequestBody));
     }
 
     @Test
     void testPostDraftAccount_trap408Response() throws Exception {
-        String validRequestBody = setupValidPostRequest();
+        String validRequestBody = validCreateRequestBody();
         shouldReturn408WhenTimeout(
             post(URL_BASE).contentType(MediaType.APPLICATION_JSON).content(validRequestBody),
             when(draftAccountService.submitDraftAccount(any())));
@@ -388,19 +94,44 @@ class DraftAccountControllerErrorsIntegrationTest {
 
     @Test
     void testPostDraftAccount_trap503Response() throws Exception {
-        String validRequestBody = setupValidPostRequest();
+        String validRequestBody = validCreateRequestBody();
         shouldReturn503WhenDownstreamServiceIsUnavailable(
             post(URL_BASE).contentType(MediaType.APPLICATION_JSON).content(validRequestBody),
-            when(draftAccountService.submitDraftAccount(any())));
+            when(draftAccountService.submitDraftAccount(any()))
+        );
+
     }
 
-    private String setupValidPostRequest() {
-        String validRequestBody = validCreateRequestBody();
-        AddDraftAccountRequestDto dto = ToJsonString.toClassInstance(validRequestBody, AddDraftAccountRequestDto.class);
-        LocalDateTime created = LocalDateTime.now();
-        DraftAccountEntity entity = toEntity(dto, created);
-        when(draftAccountService.submitDraftAccount(any())).thenReturn(entity);
-        return validRequestBody;
+    @Test
+    void testGetDraftAccountById_trap408Response() throws Exception {
+        shouldReturn408WhenTimeout(get(URL_BASE + "/1"),
+                                   when(draftAccountService.getDraftAccount(1L)));
+    }
+
+    @Test
+    void testGetDraftAccountById_trap503Response() throws Exception {
+        shouldReturn503WhenDownstreamServiceIsUnavailable(get(URL_BASE + "/1"),
+                                                          when(draftAccountService.getDraftAccount(1L)));
+    }
+
+
+
+    @Test
+    void testGetDraftAccountsSummaries_trap406Response() throws Exception {
+        shouldReturn406WhenResponseContentTypeNotSupported(get(URL_BASE));
+    }
+
+    @Test
+    void testGetDraftAccountsSummaries_trap408Response() throws Exception {
+        shouldReturn408WhenTimeout(get(URL_BASE), when(draftAccountService.getDraftAccounts(any(), any(), any(),
+                                                                                            any())));
+    }
+
+    @Test
+    void testGetDraftAccountsSummaries_trap503Response() throws Exception {
+        shouldReturn503WhenDownstreamServiceIsUnavailable(get(URL_BASE),
+                                                          when(draftAccountService.getDraftAccounts(any(), any(), any(),
+                                                                                                    any())));
     }
 
     void shouldReturn406WhenResponseContentTypeNotSupported(MockHttpServletRequestBuilder reqBuilder) throws Exception {
@@ -450,6 +181,7 @@ class DraftAccountControllerErrorsIntegrationTest {
                                               "message": "Opal Fines Database is currently unavailable"
                                           }"""));
     }
+
 
     private static String validCreateRequestBody() {
         return """
@@ -677,131 +409,19 @@ class DraftAccountControllerErrorsIntegrationTest {
             }""";
     }
 
-    private DraftAccountEntity toEntity(AddDraftAccountRequestDto dto,  LocalDateTime created) {
+    private DraftAccountEntity createDraftAccountEntity(String submittedBy, short businessUnit) {
         return DraftAccountEntity.builder()
-            .businessUnit(BusinessUnitEntity.builder().build())
-            .createdDate(created)
-            .submittedBy(dto.getSubmittedBy())
-            .account(dto.getAccount())
-            .accountSnapshot("{ \"data\": \"something snappy\"}")
-            .accountType(dto.getAccountType())
+            .draftAccountId(1L)
+            .businessUnit(BusinessUnitEntity.builder().businessUnitId(businessUnit).build())
+            .createdDate(LocalDate.of(2023, 1, 2).atStartOfDay())
+            .submittedBy(submittedBy)
+            .accountType("DRAFT")
             .accountStatus(SUBMITTED)
-            .timelineData(dto.getTimelineData())
+            .statusMessage("Status is OK")
+            .accountStatusDate(LocalDateTime.of(2024, 11, 11, 11, 11))
+            .account(validAccountJson())
+            .accountSnapshot("{ \"data\": \"something snappy\"}")
+            .timelineData(validTimelineDataJson())
             .build();
     }
-
-    //CEP 1 CEP1 - Invalid Request Payload (400)
-    @ParameterizedTest
-    @MethodSource("endpointsWithInvalidBodiesProvider")
-    void methodsShouldReturn400_whenRequestPayloadIsInvalid(
-        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
-
-        mockMvc.perform(requestBuilder
-                            .header("Authorization", "Bearer some_value")
-                            .header("Accept", "application/json")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(requestBody))
-                            .andExpect(status().isBadRequest());
-    }
-
-    private static Stream<Arguments> endpointsWithInvalidBodiesProvider() {
-        return Stream.of(Arguments.of(post(URL_BASE), invalidCreateRequestBody()),
-                         Arguments.of(put(URL_BASE + "/1"), invalidCreateRequestBody()),
-                         Arguments.of(patch(URL_BASE + "/1"), invalidCreateRequestBody())
-        );
-    }
-
-    //CEP2 - Invalid or No Access Token (401) - Security Context required - test elsewhere
-
-    //CEP3 - Not Authorised to perform the requested action (403)
-    @ParameterizedTest
-    @MethodSource("testCasesRequiringAuthorizationProvider")
-    void methodsShouldReturn403_whenUserLacksPermission(
-        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
-
-        mockMvc.perform(requestBuilder
-                            .header("Authorization", "Bearer some_value")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(requestBody))
-            .andExpect(status().isForbidden())
-            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.error").value("Forbidden"));
-    }
-
-    private static Stream<Arguments> testCasesRequiringAuthorizationProvider() {
-        return Stream.of(
-            Arguments.of(post(URL_BASE), validCreateRequestBody()),
-            Arguments.of(put(URL_BASE + "/1"), validCreateRequestBody()),
-            Arguments.of(patch(URL_BASE + "/1"), validUpdateRequestBody()),
-            Arguments.of(get(URL_BASE), "")  // GET endpoints with empty body
-        );
-    }
-
-    //CEP4 - Resource Not Found (404) - applies to GET PUT PATCH & DELETE
-    @ParameterizedTest
-    @MethodSource("testCasesForResourceNotFoundProvider")
-    void methodsShouldReturn404_whenResourceNotFound(
-        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
-        // Set up a non-existent ID
-        long nonExistentId = 999L;
-
-        // Mock the service behavior
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
-
-        // For GET return null
-        DraftAccountEntity entity = Mockito.mock(DraftAccountEntity.class);
-        when(entity.getBusinessUnit()).thenThrow(new EntityNotFoundException());
-        when(draftAccountService.getDraftAccount(nonExistentId)).thenReturn(entity);
-
-        // For PUT, throw EntityNotFoundException
-        when(draftAccountService.replaceDraftAccount(eq(nonExistentId), any(ReplaceDraftAccountRequestDto.class)))
-            .thenThrow(new jakarta.persistence.EntityNotFoundException("Draft Account not found with id: "
-                                                                           + nonExistentId));
-        // For PATCH, throw EntityNotFoundException
-        when(draftAccountService.updateDraftAccount(eq(nonExistentId), any(UpdateDraftAccountRequestDto.class)))
-            .thenThrow(new jakarta.persistence.EntityNotFoundException("Draft Account not found with id: "
-                                                                           + nonExistentId));
-        mockMvc.perform(requestBuilder
-                            .header("Authorization", "Bearer some_value")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(requestBody))
-            .andExpect(status().isNotFound());
-    }
-
-    private static Stream<Arguments> testCasesForResourceNotFoundProvider() {
-        return Stream.of(
-            Arguments.of(get(URL_BASE + "/999"), ""),
-            Arguments.of(put(URL_BASE + "/999"), validCreateRequestBody()),
-            Arguments.of(patch(URL_BASE + "/999"), validUpdateRequestBody())
-        );
-    }
-
-    //CEP5 - Unsupported Content Type for Response (406)
-    @ParameterizedTest
-    @MethodSource("testCasesWithValidBodiesProvider")
-    void methodsShouldReturn406_whenAcceptHeaderIsNotSupported(
-        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
-
-        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
-        mockMvc.perform(requestBuilder
-                            .header("Authorization", "Bearer some_value")
-                            .header("Accept", "application/xml")
-                            .contentType(MediaType.APPLICATION_JSON)
-                            .content(requestBody))
-                            .andExpect(status().isNotAcceptable());
-    }
-
-    private static Stream<Arguments> testCasesWithValidBodiesProvider() {
-        return Stream.of(Arguments.of(post(URL_BASE), validCreateRequestBody()),
-            Arguments.of(put(URL_BASE + "/1"), "{}"),
-            Arguments.of(patch(URL_BASE + "/1"), "{}"),
-            Arguments.of(get(URL_BASE + "/1"), "{}")
-        );
-    }
-
-
 }
