+++ src/integrationTest/java/uk/gov/hmcts/opal/controllers/DraftAccountControllerIntegrationTest.java
@@ -3,11 +3,15 @@ package uk.gov.hmcts.opal.controllers;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.web.servlet.MvcResult;
+import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import uk.gov.hmcts.opal.AbstractIntegrationTest;
 import uk.gov.hmcts.opal.authentication.aspect.LogAuditDetailsAspect;
 import uk.gov.hmcts.opal.authentication.service.AccessTokenService;
@@ -17,6 +21,9 @@ import uk.gov.hmcts.opal.dto.ToJsonString;
 import uk.gov.hmcts.opal.service.opal.JsonSchemaValidationService;
 import uk.gov.hmcts.opal.service.opal.UserStateService;
 
+import java.util.stream.Stream;
+
+import static org.hamcrest.Matchers.containsString;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
@@ -30,6 +37,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static uk.gov.hmcts.opal.controllers.util.UserStateUtil.allPermissionsUser;
+import static uk.gov.hmcts.opal.controllers.util.UserStateUtil.noPermissionsUser;
 import static uk.gov.hmcts.opal.controllers.util.UserStateUtil.permissionUser;
 import static uk.gov.hmcts.opal.entity.DraftAccountStatus.ERROR_IN_PUBLISHING;
 
@@ -54,6 +62,7 @@ class DraftAccountControllerIntegrationTest extends AbstractIntegrationTest {
 
     @SpyBean
     private JsonSchemaValidationService jsonSchemaValidationService;
+
 
     @Test
     void testGetDraftAccountById_success() throws Exception {
@@ -391,6 +400,373 @@ class DraftAccountControllerIntegrationTest extends AbstractIntegrationTest {
         assertTrue(jsonSchemaValidationService.isValid(body, GET_DRAFT_ACCOUNT_RESPONSE));
     }
 
+    @Test
+    void testGetDraftAccountById_trap403Response_wrongPermission() throws Exception {
+
+        UserState userState = permissionUser(BU_ID, Permissions.COLLECTION_ORDER);
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(userState);
+
+        mockMvc.perform(
+                get(URL_BASE + "/2")
+                    .header("authorization", "Bearer some_value"))
+            .andExpect(status().isForbidden())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.error").value("Forbidden"))
+            .andExpect(jsonPath("$.message").value(
+                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS, CHECK_VALIDATE_DRAFT_ACCOUNTS] "
+                    + "permission(s) are not enabled for the user."));
+    }
+
+    @Test
+    void testGetDraftAccountById_trap403Response_wrongBusinessUnit() throws Exception {
+
+        UserState userState = permissionUser((short)005, Permissions.DRAFT_ACCOUNT_PERMISSIONS);
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(userState);
+
+        mockMvc.perform(
+                get(URL_BASE + "/2")
+                    .header("authorization", "Bearer some_value"))
+            .andExpect(status().isForbidden())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.error").value("Forbidden"))
+            .andExpect(jsonPath("$.message").value(
+                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS, CHECK_VALIDATE_DRAFT_ACCOUNTS] "
+                    + "permission(s) are not enabled in business unit: 77"));
+    }
+
+    @Test
+    void testGetDraftAccountsSummaries_trap403Response_noPermission() throws Exception {
+        final Short businessId = (short)1;
+
+        UserState user = permissionUser(businessId, Permissions.COLLECTION_ORDER);
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(user);
+
+        mockMvc.perform(get(URL_BASE)
+                            .header("authorization", "Bearer some_value")
+                            .param("business_unit", businessId.toString())
+                            .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isForbidden())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON));
+    }
+
+    @Test
+    void testGetDraftAccountById_trap404Response() throws Exception {
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
+
+        mockMvc.perform(
+                get(URL_BASE + "/99")
+                    .header("authorization", "Bearer some_value"))
+            .andExpect(status().isNotFound())
+        ;
+    }
+
+    @Test
+    void testGetDraftAccountById_trap406Response() throws Exception {
+        mockMvc.perform(get(URL_BASE + "/99")
+                            .header("Authorization", "Bearer " + "some_value")
+                            .accept("application/xml"))
+            .andExpect(status().isNotAcceptable());
+    }
+
+    @Test
+    void testGetDraftAccountsSummaries_trap400Response() throws Exception {
+        final Short businessId = (short)1;
+
+        UserState user = permissionUser(businessId, Permissions.CREATE_MANAGE_DRAFT_ACCOUNTS);
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(user);
+
+        mockMvc.perform(get(URL_BASE)
+                            .header("authorization", "Bearer some_value")
+                            .param("business_unit", businessId.toString())
+                            .param("submitted_by", "Dave")
+                            .param("not_submitted_by", "Tony")
+                            .contentType(MediaType.APPLICATION_JSON))
+            .andExpect(status().isBadRequest())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.error").value("Bad Request"))
+            .andExpect(jsonPath("$.message")
+                           .value("Cannot include both 'submitted_by' and 'not_submitted_by' parameters."));
+    }
+
+    @Test
+    void testReplaceDraftAccount_trap403Response_noPermission() throws Exception {
+        Long draftAccountId = 241L;
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
+
+        mockMvc.perform(put(URL_BASE + "/" + draftAccountId)
+                            .header("authorization", "Bearer some_value")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(validCreateRequestBody()))
+            .andExpect(status().isForbidden())
+            .andReturn();
+
+    }
+
+    @Test
+    void testUpdateDraftAccount_trap403Response_noPermission() throws Exception {
+        Long draftAccountId = 241L;
+        String requestBody = "            {\n"
+            + "                \"account_status\": \"PENDING\",\n"
+            + "                \"validated_by\": \"BUUID1\",\n"
+            + "                \"business_unit_id\": 5,\n"
+            + "                \"timeline_data\": "
+            + validTimelineDataJson()
+            + "\n"
+            + "            }";
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
+
+        mockMvc.perform(patch(URL_BASE + "/" + draftAccountId)
+                            .header("authorization", "Bearer some_value")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(requestBody))
+            .andExpect(status().isForbidden())
+            .andReturn();
+
+    }
+
+    @Test
+    void testPostDraftAccount_trap400Response() throws Exception {
+
+        String expectedErrorMessageStart =
+            "JSON Schema Validation Error: Validating against JSON schema 'addDraftAccountRequest.json',"
+                + " found 14 validation errors:";
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
+
+        mockMvc.perform(post(URL_BASE)
+                            .header("authorization", "Bearer some_value")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(invalidCreateRequestBody()))
+            .andExpect(status().isBadRequest())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.error").value("Bad Request"))
+            .andExpect(jsonPath("$.message").value(containsString(expectedErrorMessageStart)))
+            .andExpect(jsonPath("$.message").value(containsString("required property 'account_type' not found")))
+            .andExpect(jsonPath("$.message").value(containsString("required property 'submitted_by' not found")))
+            .andExpect(jsonPath("$.message").value(containsString("required property 'submitted_by_name' not found")))
+            .andExpect(jsonPath("$.message").value(containsString("required property 'timeline_data' not found")));
+
+    }
+
+    @Test
+    void testPostDraftAccount_trap403Response_noPermission() throws Exception {
+
+        String validRequestBody = validCreateRequestBody();
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
+
+        MvcResult result = mockMvc.perform(post(URL_BASE)
+                                               .header("authorization", "Bearer some_value")
+                                               .contentType(MediaType.APPLICATION_JSON)
+                                               .content(validRequestBody))
+            .andExpect(status().isForbidden())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.error").value("Forbidden"))
+            .andExpect(jsonPath("$.message").value(
+                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS] permission(s) are not enabled for the user."))
+            .andReturn();
+
+        String body = result.getResponse().getContentAsString();
+
+        log.info(":testPostDraftAccount_permission: Response body:\n" + ToJsonString.toPrettyJson(body));
+    }
+
+    @Test
+    void testPostDraftAccount_trap403Response_wrongPermission() throws Exception {
+
+        String validRequestBody = validCreateRequestBody();
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(
+            permissionUser((short)5, Permissions.CHECK_VALIDATE_DRAFT_ACCOUNTS, Permissions.ACCOUNT_ENQUIRY));
+
+        MvcResult result = mockMvc.perform(post(URL_BASE)
+                                               .header("authorization", "Bearer some_value")
+                                               .contentType(MediaType.APPLICATION_JSON)
+                                               .content(validRequestBody))
+            .andExpect(status().isForbidden())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.error").value("Forbidden"))
+            .andExpect(jsonPath("$.message").value(
+                "For user null, [CREATE_MANAGE_DRAFT_ACCOUNTS] permission(s) are not enabled for the user."))
+            .andReturn();
+
+        String body = result.getResponse().getContentAsString();
+
+        log.info(":testPostDraftAccount_permission: Response body:\n" + ToJsonString.toPrettyJson(body));
+    }
+
+    //CEP 1 CEP1 - Invalid Request Payload (400)
+    @ParameterizedTest
+    @MethodSource("endpointsWithInvalidBodiesProvider")
+    void methodsShouldReturn400_whenRequestPayloadIsInvalid(
+        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
+
+        mockMvc.perform(requestBuilder
+                            .header("Authorization", "Bearer some_value")
+                            .header("Accept", "application/json")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(requestBody))
+            .andExpect(status().isBadRequest());
+    }
+
+    private static Stream<Arguments> endpointsWithInvalidBodiesProvider() {
+        return Stream.of(Arguments.of(post(URL_BASE), invalidCreateRequestBody()),
+                         Arguments.of(put(URL_BASE + "/1"), invalidCreateRequestBody()),
+                         Arguments.of(patch(URL_BASE + "/1"), invalidCreateRequestBody())
+        );
+    }
+
+    //CEP3 - Not Authorised to perform the requested action (403)
+    @ParameterizedTest
+    @MethodSource("testCasesRequiringAuthorizationProvider")
+    void methodsShouldReturn403_whenUserLacksPermission(
+        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(noPermissionsUser());
+
+        mockMvc.perform(requestBuilder
+                            .header("Authorization", "Bearer some_value")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(requestBody))
+            .andExpect(status().isForbidden())
+            .andExpect(content().contentType(MediaType.APPLICATION_JSON))
+            .andExpect(jsonPath("$.error").value("Forbidden"));
+    }
+
+    private static Stream<Arguments> testCasesRequiringAuthorizationProvider() {
+        return Stream.of(
+            Arguments.of(post(URL_BASE), validCreateRequestBody()),
+            Arguments.of(put(URL_BASE + "/1"), validCreateRequestBody()),
+            Arguments.of(patch(URL_BASE + "/1"), validUpdateRequestBody()),
+            Arguments.of(get(URL_BASE), "")  // GET endpoints with empty body
+        );
+    }
+
+    //CEP4 - Resource Not Found (404) - applies to GET PUT PATCH & DELETE
+    @ParameterizedTest
+    @MethodSource("testCasesForResourceNotFoundProvider")
+    void methodsShouldReturn404_whenResourceNotFound(
+        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
+        // Set up a non-existent ID
+        long nonExistentId = 999L;
+
+        // Mock the service behavior
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
+
+        mockMvc.perform(requestBuilder
+                            .header("Authorization", "Bearer some_value")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(requestBody))
+            .andExpect(status().isNotFound());
+    }
+
+    private static Stream<Arguments> testCasesForResourceNotFoundProvider() {
+        return Stream.of(
+            Arguments.of(get(URL_BASE + "/999"), ""),
+            Arguments.of(put(URL_BASE + "/999"), validCreateRequestBody()),
+            Arguments.of(patch(URL_BASE + "/999"), validUpdateRequestBody())
+        );
+    }
+
+    //CEP5 - Unsupported Content Type for Response (406)
+    @ParameterizedTest
+    @MethodSource("testCasesWithValidBodiesProvider")
+    void methodsShouldReturn406_whenAcceptHeaderIsNotSupported(
+        MockHttpServletRequestBuilder requestBuilder, String requestBody) throws Exception {
+
+        when(userStateService.checkForAuthorisedUser(any())).thenReturn(allPermissionsUser());
+        mockMvc.perform(requestBuilder
+                            .header("Authorization", "Bearer some_value")
+                            .header("Accept", "application/xml")
+                            .contentType(MediaType.APPLICATION_JSON)
+                            .content(requestBody))
+            .andExpect(status().isNotAcceptable());
+    }
+
+    private static Stream<Arguments> testCasesWithValidBodiesProvider() {
+        return Stream.of(Arguments.of(post(URL_BASE), validCreateRequestBody()),
+                         Arguments.of(put(URL_BASE + "/1"), "{}"),
+                         Arguments.of(patch(URL_BASE + "/1"), "{}"),
+                         Arguments.of(get(URL_BASE + "/1"), "{}")
+        );
+    }
+
+    private final String validAccountJson() {
+        return """
+            {
+              "account_type": "Fine",
+              "defendant_type": "Adult",
+              "originator_name": "Police Force",
+              "originator_id": 12345,
+              "enforcement_court_id": 101,
+              "collection_order_made": true,
+              "collection_order_made_today": false,
+              "payment_card_request": true,
+              "account_sentence_date": "2023-12-01",
+              "defendant": {
+                "company_flag": true,
+                "company_name": "company",
+                "dob": "1985-04-15",
+                "address_line_1": "123 Elm Street",
+                "address_line_2": "Suite 45",
+                "post_code": "AB1 2CD",
+                "telephone_number_home": "0123456789",
+                "telephone_number_mobile": "07712345678",
+                "email_address_1": "john.doe@example.com",
+                "national_insurance_number": "AB123456C",
+                "nationality_1": "British",
+                "occupation": "Engineer",
+                "debtor_detail": {
+                  "document_language": "English",
+                  "hearing_language": "English",
+                  "vehicle_make": "Toyota",
+                  "vehicle_registration_mark": "ABC123",
+                  "aliases": [
+                    {
+                      "alias_forenames": "Jon",
+                      "alias_surname": "Smith"
+                    }
+                  ]
+                }
+              },
+              "offences": [
+                {
+                  "date_of_sentence": "2023-11-15",
+                  "imposing_court_id": 202,
+                  "offence_id": 1234,
+                  "impositions": [
+                    {
+                      "result_id": "1",
+                      "amount_imposed": 500.00,
+                      "amount_paid": 200.00,
+                      "major_creditor_id": 999
+                    }
+                  ]
+                }
+              ],
+              "payment_terms": {
+                "payment_terms_type_code": "P",
+                "effective_date": "2023-11-01",
+                "instalment_period": "M",
+                "lump_sum_amount": 1000.00,
+                "instalment_amount": 200.00,
+                "default_days_in_jail": 5
+              },
+              "account_notes": [
+                {
+                  "account_note_serial": 1,
+                  "account_note_text": "Defendant requested an installment plan.",
+                  "note_type": "AC"
+                }
+              ]
+            }""";
+    }
+
     private static String validCreateRequestBody() {
         return """
             {
@@ -523,6 +899,27 @@ class DraftAccountControllerIntegrationTest extends AbstractIntegrationTest {
                     "reason_text": "Violation of terms of service."
                 }
             ]""";
+    }
+
+    private static String invalidCreateRequestBody() {
+        return """
+            {
+             "invalid_field": "This field shouldn't be here",
+             "account": {
+                "account_create_request": {
+                "defendant": {
+                    "company_name": "Company ABC",
+                    "surname": "LNAME",
+                    "fornames": "FNAME",
+                    "dob": "2000-01-01"
+                },
+             "account": {
+                "account_type": "Invalid"
+             }
+            }
+            },
+                "business_unit_id": 1
+            }""";
     }
 
 }
